delete: 컨트롤러 인터페이스 삭제
